requirement.txt
crewai[tools]
requests
python-dotenv
streamlit